Refactor log level setting into separate utility function
This commit primarily refactors the setting of the log level into a separate function located in the logger module. As part of this change, the logger's exports were updated to be an object that includes both the logger itself and the new setLogLevel function. Additional changes include replacing every import of the logger to account for the new export structure.