Make Docker documentation base image agnostic
Not all Docker images come with a shell. Notably, Scratch and Distroless come without a shell to reduce the image size and harden the image. If you don't have a shell, it can't be exploited.

The easiest solution is to simply copy the binaries for the official image. Since buildx will create an image for each architecture, we don't need an architecture-specific command.

Additionally, it's best to use fully resolved entry point paths just in case someone changes the $PATH variable.

Close #713